Install initial domain mapping in metadata extraction phase, add tests for state (#64)
## Summary
<!-- Provide a brief description of the story behind this PR, as if
explaining
     to a non-technical person. Or to an LLM so it can learn from it for
future (autonomous) code improvements. Feel free to point to a deeper
     design doc, if applicable.
-->
This PR adds `ExtractionMetadataStart` event to list of stateful event
types. This way installation of initial domain mappings happens in
metadata phase too. Besides that, state implementation code is
refactored for readability and added a lot of behavior tests for state.

## Connected Issues
<!-- Have you cared to connect this PR to a work item in DevRev, so that
we
     can understand future routing and attribution?
-->
- https://app.devrev.ai/devrev/works/ISS-205447
- https://app.devrev.ai/devrev/works/ISS-205447